Refaktorerte ut datatransformasjoner fra Infographics-komponenten
