+++ client/src/Infographic/Infographic.tsx
@@ -12,43 +12,41 @@ import { HealthCase } from "@navikt/ds-icons";
 import { Up } from "@navikt/ds-icons";
 import { BodyLong, Link } from "@navikt/ds-react";
 
+export interface InfographicData {
+  sykefraværNorge: number | null | undefined,
+  sykefraværNæring: number | null | undefined,
+  trendStigningstall: number | undefined,
+}
+
 export const Infographic: FunctionComponent<{
-  historikk: KvartalsvisSykefraværshistorikk[];
-}> = ({ historikk }) => {
+  data: InfographicData;
+}> = ({ data }) => {
   const ikonstorrelse = { width: '50px', height: '50px' };
-  let sisteFraværsprosentNorge = sisteTilgjengelige(
-    sykefraværINorge(historikk)
-  )?.prosent;
 
-  let sisteFraværsprosentNæring = sisteTilgjengelige(
-    getSykefraværForNæring(historikk)
-  )?.prosent;
-
-  let trend = kalkulerTrend(getSykefraværForNæring(historikk));
-  let trendSomTekst = stigningskurveTilTekst(trend);
+  let trendSomTekst = stigningskurveTilTekst(data.trendStigningstall);
 
   return (
     <div className={styles.infographicWrapper}>
       <InfographicSnippet
-        ikon={<NorwegianFlag width={ikonstorrelse.width} height={ikonstorrelse.height} />}
+        ikon={<NorwegianFlag {...ikonstorrelse} />}
         tekst={"Sykefraværsprosenten i Norge akkurat nå er: "}
-        verdi={sisteFraværsprosentNorge + "%"}
+        verdi={data.sykefraværNorge + "%"}
       />
 
       <InfographicSnippet
-        ikon={<Bag width={ikonstorrelse.width} height={ikonstorrelse.height} />}
+        ikon={<Bag {...ikonstorrelse} />}
         tekst={"Sykefraværsprosenten i din bransje akkurat nå er: "}
-        verdi={sisteFraværsprosentNæring + "%"}
+        verdi={data.sykefraværNæring + "%"}
       />
 
       <InfographicSnippet
-        ikon={<HealthCase width={ikonstorrelse.width} height={ikonstorrelse.height} />}
+        ikon={<HealthCase {...ikonstorrelse} />}
         tekst={"Vanligste årsak til sykemelding i Norge er: "}
         verdi={"Muskel- og skjelettplager"}
       />
 
       <InfographicSnippet
-        ikon={<Up width={ikonstorrelse.width} height={ikonstorrelse.height} />}
+        ikon={<Up {...ikonstorrelse} />}
         tekst={"Sykefraværet i din bransje akkurat nå er "}
         verdi={trendSomTekst}
       />
@@ -63,76 +61,7 @@ export const Infographic: FunctionComponent<{
   );
 };
 
-// TODO: Refaktorer ut disse funksjonene. Denne komponenten skal ikke være ansvarlig for datatransformasjon.
-// Ideelt skal ikke mer data eksponeres til komponenten enn akkurat hva den har bruk for.
-function sykefraværINorge(
-  fraværshistorikk: KvartalsvisSykefraværshistorikk[]
-): KvartalsvisSykefraværshistorikk {
-  const [historikkForNorge] = fraværshistorikk.filter(
-    (historikk) =>
-      historikk.type === SykefraværshistorikkType.LAND &&
-      historikk.label === "Norge"
-  );
 
-  return historikkForNorge;
-}
-
-function sisteTilgjengelige(
-  historikk: KvartalsvisSykefraværshistorikk
-): KvartalsvisSykefraværsprosent | null {
-  if (!historikk?.kvartalsvisSykefraværsprosent) {
-    return null;
-  }
-  const sortert = sorterKronologisk(historikk);
-  return sortert[sortert.length - 1];
-}
-
-function getSykefraværForNæring(
-  fraværshistorikk: KvartalsvisSykefraværshistorikk[]
-): KvartalsvisSykefraværshistorikk {
-  const [historikkForNæring] = fraværshistorikk.filter(
-    (historikk) => historikk.type === SykefraværshistorikkType.NÆRING
-  );
-
-  return historikkForNæring;
-}
-
-function getDataForEnkeltkvartal(
-  fraværshistorikk: KvartalsvisSykefraværsprosent[],
-  årstall: number,
-  kvartal: number
-): KvartalsvisSykefraværsprosent {
-  const [enkeltkvartal] = fraværshistorikk.filter(
-    (kvartalstall) =>
-      kvartalstall.årstall === årstall && kvartalstall.kvartal === kvartal
-  );
-
-  return enkeltkvartal;
-}
-
-function kalkulerTrend(
-  fraværsistorikk: KvartalsvisSykefraværshistorikk
-): number | undefined {
-  if (!fraværsistorikk?.kvartalsvisSykefraværsprosent) {
-    return undefined;
-  }
-  const sortert = sorterKronologisk(fraværsistorikk);
-  const nyesteDatapunkt = sortert[sortert.length - 1];
-  const sisteTilgjengeligeÅrstall = nyesteDatapunkt.årstall;
-  const sisteTilgjengeligeKvartal = nyesteDatapunkt.kvartal;
-  const prosentSisteKvartal = nyesteDatapunkt.prosent;
-
-  const prosentSammeKvartalIFjor = getDataForEnkeltkvartal(
-    fraværsistorikk.kvartalsvisSykefraværsprosent,
-    sisteTilgjengeligeÅrstall - 1,
-    sisteTilgjengeligeKvartal
-  )?.prosent;
-
-  if (prosentSisteKvartal && prosentSammeKvartalIFjor) {
-    return prosentSisteKvartal - prosentSammeKvartalIFjor;
-  }
-  return undefined;
-}
 
 function stigningskurveTilTekst(stigning: number | undefined): string {
   if (stigning === undefined) {
@@ -144,10 +73,4 @@ function stigningskurveTilTekst(stigning: number | undefined): string {
   } else {
     return "uendret";
   }
-}
-
-function sorterKronologisk(historikk: KvartalsvisSykefraværshistorikk) {
-  return historikk.kvartalsvisSykefraværsprosent.sort(
-    (e1, e2) => e1.årstall - e2.årstall || e1.kvartal - e2.kvartal
-  );
 }
